# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: goals.ac API — Programmatic SEO Growth Roadmap Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /content-strategies/generate:
    post:
      operationId: generateContentStrategy
      tags: [content-strategies]
      summary: Generate a 30-day content strategy
      description: Generates a new 30-day content strategy using AI for the given roadmap parameters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateContentStrategyRequest"
      responses:
        "200":
          description: Content strategy generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentStrategyWithItems"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Roadmap not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /content-strategies/{id}:
    get:
      operationId: getContentStrategy
      tags: [content-strategies]
      summary: Get a content strategy with all items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content strategy with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentStrategyWithItems"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /content-strategies/{id}/items/{itemId}:
    patch:
      operationId: updateContentItem
      tags: [content-strategies]
      summary: Update a content item status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContentItemRequest"
      responses:
        "200":
          description: Content item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /competitor-analysis:
    post:
      operationId: createCompetitorAnalysis
      tags: [seo-tools]
      summary: Analyze a competitor URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompetitorAnalysisRequest"
      responses:
        "201":
          description: Competitor analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitorAnalysisResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: AI unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /competitor-analyses/{id}:
    get:
      operationId: getCompetitorAnalysis
      tags: [seo-tools]
      summary: Get a saved competitor analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Competitor analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitorAnalysisResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    GenerateContentStrategyRequest:
      type: object
      properties:
        roadmap_id:
          type: integer
        industry:
          type: string
        location:
          type: string
        stage:
          type: string
      required:
        - roadmap_id
        - industry
        - location
        - stage
    UpdateContentItemRequest:
      type: object
      properties:
        status:
          type: string
          enum: [draft, prepared, published]
      required:
        - status
    ContentItem:
      type: object
      properties:
        id:
          type: integer
        strategyId:
          type: integer
        day:
          type: integer
        title:
          type: string
        format:
          type: string
        topicAngle:
          type: string
        primaryKeyword:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - strategyId
        - day
        - title
        - format
        - topicAngle
        - primaryKeyword
        - status
        - createdAt
...
    ContentStrategyWithItems:
      type: object
      properties:
        id:
          type: integer
        roadmapId:
          type: integer
        industry:
          type: string
        location:
          type: string
        stage:
          type: string
        month:
          type: integer
        year:
          type: integer
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContentItem"
      required:
        - id
        - roadmapId
        - industry
        - location
        - stage
        - month
        - year
        - createdAt
        - items
...
    CreateCompetitorAnalysisRequest:
      type: object
      properties:
        competitorUrl:
          type: string
          format: uri
        industry:
          type: string
        location:
          type: string
        stage:
          type: string
        website_project_id:
          type: integer
      required: [competitorUrl, industry, location, stage]
    CompetitorAnalysisResponse:
      type: object
      properties:
        id:
          type: integer
        competitorName:
          type: string
        summary:
          type: string
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        contentGaps:
          type: array
          items:
            type: string
        geoGaps:
          type: array
          items:
            type: string
        quickWins:
          type: array
          items:
            type: string
        threatLevel:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
          format: date-time
      required:
        - competitorName
        - summary
        - strengths
        - weaknesses
        - contentGaps
        - geoGaps
        - quickWins
        - threatLevel
...